+++ src/Sfa.Tl.Matching.Application.UnitTests/Services/ProviderFeedback/When_ProviderQuarterlyUpdateEmailService_Is_Called_To_Send_Provider_Quarterly_Update_Emails_With_No_Qualifications.cs
@@ -3,9 +3,7 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Linq.Expressions;
 using FluentAssertions;
-using Microsoft.Extensions.Logging;
 using NSubstitute;
-using NSubstitute.ExceptionExtensions;
 using Sfa.Tl.Matching.Application.Interfaces;
 using Sfa.Tl.Matching.Application.Services;
 using Sfa.Tl.Matching.Application.UnitTests.Services.ProviderFeedback.Builders;
@@ -16,35 +14,36 @@ using Xunit;
 
 namespace Sfa.Tl.Matching.Application.UnitTests.Services.ProviderFeedback
 {
-    public class When_ProviderFeedbackService_Send_Provider_Quarterly_Update_Emails_Throws_Exception
-        : IClassFixture<ProviderFeedbackFixture>
+    public class When_ProviderQuarterlyUpdateEmailService_Is_Called_To_Send_Provider_Quarterly_Update_Emails_With_No_Qualifications
+        : IClassFixture<ProviderQuarterlyUpdateEmailFixture>
     {
+        private readonly ProviderQuarterlyUpdateEmailFixture _testFixture;
         private readonly IEmailService _emailService;
         private readonly IEmailHistoryService _emailHistoryService;
         private readonly IProviderRepository _providerRepository;
         private readonly IRepository<BackgroundProcessHistory> _backgroundProcessHistoryRepository;
         private readonly IList<BackgroundProcessHistory> _receivedProviderFeedbackRequestHistories;
-        private readonly ILogger<ProviderFeedbackService> _logger;
         private readonly int _result;
 
-        public When_ProviderFeedbackService_Send_Provider_Quarterly_Update_Emails_Throws_Exception(ProviderFeedbackFixture testFixture)
+        public When_ProviderQuarterlyUpdateEmailService_Is_Called_To_Send_Provider_Quarterly_Update_Emails_With_No_Qualifications(ProviderQuarterlyUpdateEmailFixture testFixture)
         {
+            _testFixture = testFixture;
+
             _emailService = Substitute.For<IEmailService>();
             _emailHistoryService = Substitute.For<IEmailHistoryService>();
 
             var messageQueueService = Substitute.For<IMessageQueueService>();
-            _logger = Substitute.For<ILogger<ProviderFeedbackService>>();
-
+           
             _providerRepository = Substitute.For<IProviderRepository>();
             _providerRepository
                 .GetProvidersWithFundingAsync()
-                .Returns(new ValidProviderWithFundingDtoListBuilder().Build());
+                .Returns(new ValidProviderWithFundingDtoListBuilder().BuildWithNoQualifications());
 
             _backgroundProcessHistoryRepository = Substitute.For<IRepository<BackgroundProcessHistory>>();
             _backgroundProcessHistoryRepository
                 .GetSingleOrDefault(Arg.Any<Expression<Func<BackgroundProcessHistory, bool>>>())
                 .Returns(new BackgroundProcessHistoryBuilder().Build());
-            
+
             _receivedProviderFeedbackRequestHistories = new List<BackgroundProcessHistory>();
             _backgroundProcessHistoryRepository
                 .Update(Arg.Do<BackgroundProcessHistory>
@@ -54,21 +53,16 @@ namespace Sfa.Tl.Matching.Application.UnitTests.Services.ProviderFeedback
                         Id = x.Id,
                         RecordCount = x.RecordCount,
                         Status = x.Status,
-                        StatusMessage = x.StatusMessage,
                         ModifiedOn = x.ModifiedOn,
                         ModifiedBy = x.ModifiedBy
                     }
                 )));
 
-            _emailService
-                .SendEmail(Arg.Any<string>(), Arg.Any<string>(), Arg.Any<string>(), Arg.Any<IDictionary<string, string>>(), Arg.Any<string>())
-                .Throws(new Exception());
-
-            var providerFeedbackService = new ProviderFeedbackService(
-                testFixture.Configuration, _logger, 
+            var providerFeedbackService = new ProviderQuarterlyUpdateEmailService(
+                _testFixture.Configuration, _testFixture.Logger,
                     _emailService, _emailHistoryService,
                     _providerRepository, _backgroundProcessHistoryRepository,
-                    messageQueueService, testFixture.DateTimeProvider);
+                    messageQueueService, _testFixture.DateTimeProvider);
 
             _result = providerFeedbackService
                 .SendProviderQuarterlyUpdateEmailsAsync(1, "TestUser")
@@ -90,7 +84,7 @@ namespace Sfa.Tl.Matching.Application.UnitTests.Services.ProviderFeedback
                 .Received(1)
                 .GetSingleOrDefault(Arg.Any<Expression<Func<BackgroundProcessHistory, bool>>>());
         }
-        
+
         [Fact]
         public void Then_BackgroundProcessHistoryRepository_Update_Is_Called_Exactly_Twice()
         {
@@ -102,35 +96,33 @@ namespace Sfa.Tl.Matching.Application.UnitTests.Services.ProviderFeedback
         [Fact]
         public void Then_BackgroundProcessHistoryRepository_Update_Is_Called_With_Expected_Values()
         {
-            //Can't check Status and count here because NSubstitute only remembers the last one
+            //Can't check Status here because NSubstitute only remembers the last one
             _backgroundProcessHistoryRepository
                 .Received()
                 .Update(Arg.Is<BackgroundProcessHistory>(
                     p => p.Id == 1
-                         //&& p.RecordCount == 1
+                         && p.RecordCount == 1
                          && p.ModifiedBy == "TestUser"
                 ));
         }
-        
+
         [Fact]
         public void Then_BackgroundProcessHistoryRepository_Update_Sets_Expected_Values_In_First_Call()
         {
             var history = _receivedProviderFeedbackRequestHistories.First();
             history.Id.Should().Be(1);
             history.Status.Should().Be(BackgroundProcessHistoryStatus.Processing.ToString());
-            history.StatusMessage.Should().BeNullOrEmpty();
             history.RecordCount.Should().Be(1);
             history.ModifiedBy.Should().Be("TestUser");
         }
-        
+
         [Fact]
         public void Then_BackgroundProcessHistoryRepository_Update_Sets_Expected_Values_In_Second_Call()
         {
             var history = _receivedProviderFeedbackRequestHistories.Skip(1).First();
             history.Id.Should().Be(1);
-            history.Status.Should().Be(BackgroundProcessHistoryStatus.Error.ToString());
-            history.StatusMessage.Should().NotBeNullOrEmpty();
-            history.RecordCount.Should().Be(0);
+            history.Status.Should().Be(BackgroundProcessHistoryStatus.Complete.ToString());
+            history.RecordCount.Should().Be(1);
             history.ModifiedBy.Should().Be("TestUser");
         }
 
@@ -141,33 +133,65 @@ namespace Sfa.Tl.Matching.Application.UnitTests.Services.ProviderFeedback
                 .Received(1)
                 .SendEmail(Arg.Any<string>(), Arg.Any<string>(), Arg.Any<string>(), Arg.Any<IDictionary<string, string>>(), Arg.Any<string>());
         }
-
+        
         [Fact]
-        public void Then_Logger_Is_Called_With_Error_Message()
+        public void Then_EmailService_SendEmail_Is_Called_With_Expected_Parameters()
         {
-            _logger.Received(1).Log(
-                LogLevel.Error,
-                Arg.Any<EventId>(),
-                Arg.Is<object>(o =>
-                    o.ToString()
-                        .Contains(
-                            "Error sending provider quarterly update emails.")),
-                Arg.Any<Exception>(),
-                Arg.Any<Func<object, Exception, string>>());
+            _emailService
+                .Received(1)
+                .SendEmail(Arg.Is<string>(
+                        templateName => templateName == "ProviderQuarterlyUpdate"),
+                    Arg.Is<string>(toAddress => toAddress == "primary.contact@provider.co.uk"),
+                    Arg.Is<string>(subject => subject == "Industry Placement Matching Provider Update"),
+                    Arg.Any<IDictionary<string, string>>(),
+                    Arg.Is<string>(replyToAddress => replyToAddress == ""));
+        }
+        
+        [Fact]
+        public void Then_EmailService_SendEmail_Is_Called_With_Expected_Tokens()
+        {
+            const string expectedSecondaryDetailsList = "We also have the following secondary contact for Provider Name:\r\n"
+                                                        + "* Name: SecondaryContact\r\n"
+                                                        + "* Email: secondary@contact.co.uk\r\n"
+                                                        + "* Telephone: 01234559999\r\n";
+            const string expectedProviderVenueQualificationsList =
+                "AA1 1AA:\r\n"
+                + "* no qualifications with an industry placement option\r\n"
+                + "\r\n";
+
+            var expectedResults = new Dictionary<string, string>
+            {
+                { "provider_name",  "Provider Name" },
+                { "primary_contact_name", "Provider Contact" },
+                { "primary_contact_email", "primary.contact@provider.co.uk" },
+                { "primary_contact_phone", "01777757777" },
+                { "secondary_contact_details", expectedSecondaryDetailsList },
+                { "provider_has_venues", "yes" },
+                { "provider_has_no_venues", "no" },
+                { "venues_and_qualifications_list", expectedProviderVenueQualificationsList }
+            };
+
+            _emailService
+                .Received(1)
+                .SendEmail(Arg.Any<string>(),
+                    Arg.Any<string>(),
+                    Arg.Any<string>(),
+                    Arg.Is<IDictionary<string, string>>(
+                        tokens => _testFixture.DoTokensContainExpectedValues(tokens, expectedResults)),
+                    Arg.Any<string>());
         }
 
         [Fact]
-        public void Then_EmailHistoryService_SaveEmailHistory_Is_Not_Called()
+        public void Then_EmailHistoryService_SaveEmailHistory_Is_Called_Exactly_Once()
         {
             _emailHistoryService
-                .DidNotReceiveWithAnyArgs()
+                .Received(1)
                 .SaveEmailHistory(Arg.Any<string>(), Arg.Any<IDictionary<string, string>>(), Arg.Any<int?>(), Arg.Any<string>(), Arg.Any<string>());
         }
-
         [Fact]
         public void Then_Result_Has_Expected_Value()
         {
-            _result.Should().Be(0);
+            _result.Should().Be(1);
         }
     }
 }
