+++ src/Sfa.Tl.Matching.Web/Startup.cs
@@ -257,7 +257,7 @@ namespace Sfa.Tl.Matching.Web
             services.AddTransient<IRoutePathService, RoutePathService>();
             services.AddTransient<IOpportunityService, OpportunityService>();
             services.AddTransient<IProviderService, ProviderService>();
-            services.AddTransient<IProviderFeedbackService, ProviderFeedbackService>();
+            services.AddTransient<IProviderQuarterlyUpdateEmailService, ProviderQuarterlyUpdateEmailService>();
             services.AddTransient<IProximityService, ProximityService>();
             services.AddTransient<IReferralService, ReferralService>();
             services.AddTransient<IProviderVenueService, ProviderVenueService>();
