+++ src/Sfa.Tl.Matching.Web/Views/ProviderQuarterlyUpdateEmail/ConfirmSendProviderEmail.cshtml
@@ -23,7 +23,7 @@
         }
 
         <h1 class="govuk-heading-l">@ViewData["Title"]</h1>
-        <form asp-controller="ProviderFeedback" asp-action="ConfirmSendProviderEmail">
+        <form asp-controller="ProviderQuarterlyUpdateEmail" asp-action="ConfirmSendProviderEmail">
 
             <div class="govuk-form-group" sfa-validation-for="SendEmail" sfa-validationerror-class="govuk-form-group--error">
                 <span class="govuk-error-message" asp-validation-for="SendEmail"></span>
